chore: move types to single file
