+++ packages/toolkit/plugin/src/manager/index.ts
@@ -1,3 +1,4 @@
 export * from './sync';
 export * from './async';
 export * from './runner';
+export * from './types';
